Revert "move all of the localizers and then change them to implement updatePositionEstimate and updateVelocityEstimate instead of simply "update""
This reverts commit 64d825f24cf5ac7bb2c1ac5e62ce4ffd61f808f3.
